sharing types [nfc]: Move SharedData inside src/sharing/.
This type definition belongs squarely to a nice well-defined
subsystem, and we have a nice subdirectory for that subsystem.
So put the definition there, alongside the other (JS) code
belonging to that subsystem.

This is also good for reducing cycles in our type-import graph.
